# docs/swagger.yaml
info:
  contact: {}
paths:
  /search:
    get:
      consumes:
      - application/json
      description: Query Meilisearch and return results
      parameters:
      - description: name of index
        in: query
        name: index
        required: true
        type: string
      - description: Search query string
        in: query
        name: query
        required: true
        type: string
      - description: Number of results to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search in Meilisearch
      tags:
      - search
  /search/basic:
    get:
      consumes:
      - application/json
      description: Query All Index in Meilisearch and return results
      parameters:
      - description: Search query string
        in: query
        name: query
        required: true
        type: string
      - description: Number of results to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Basic Combined Search in Meilisearch
      tags:
      - search
  /search/final:
    get:
      consumes:
      - application/json
      description: Query All Index in Meilisearch and return results
      parameters:
      - description: Search query string
        in: query
        name: query
        required: true
        type: string
      - description: Number of results to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Final Combined Search in Meilisearch
      tags:
      - search
securityDefinitions:
  cookie:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"